+++ src/ziggurat/mapper.clj
@@ -4,13 +4,17 @@
             [ziggurat.messaging.producer :as producer]
             [ziggurat.sentry :refer [sentry-reporter]]
             [sentry.core :as sentry]
-            [clojure.tools.logging :as log]))
+            [clojure.tools.logging :as log])
+  (:import (java.time Instant)))
 
 (defn mapper-func [mapper-fn]
   (fn [message topic-name]
     (nr/with-tracing "job" "mapper-func"
       (try
-        (let [return-code (mapper-fn message)]
+        (let [start-time (.toEpochMilli (Instant/now))
+              return-code (mapper-fn message)
+              end-time (.toEpochMilli (Instant/now))]
+          (metrics/report-time "mapper-function-execution-time" (- end-time start-time))
           (case return-code
             :success (metrics/message-successfully-processed!)
             :retry (do (metrics/message-unsuccessfully-processed!)